+++ src/JwtWebApi.Api.Services/Dto/ObjectEntities.cs
@@ -3,6 +3,16 @@ using JwtWebApi.DataProviders.Common.DataObjects;
 
 namespace JwtWebApi.Api.Services.Dto
 {
+	public interface IRestaurantWithLinks
+	{
+		public IRestaurant Restaurant { get; set; }
+
+		public ICuisineType[] CuisineTypes { get; set; }
+
+		public IDenyType[] DenyTypes { get; set; }
+	}
+
+
 	public interface IRestaurant : IEntity
 	{
 		public string Name { get; set; }
