+++ src/JwtWebApi.Api.Services/Services/ObjectsServicesContracts.cs
@@ -1,11 +1,13 @@
-﻿using JwtWebApi.Api.Common.Services;
+﻿using System.Threading.Tasks;
+using JwtWebApi.Api.Common.Dto;
+using JwtWebApi.Api.Common.Services;
 using JwtWebApi.Api.Services.Dto;
 
 namespace JwtWebApi.Api.Services.Services
 {
 	public interface IRestaurantService : IEntityProvider<IRestaurant>
 	{
-
+		Task<PagingResult<IRestaurantWithLinks>> GetPagingWithLinks(int page, int pageSize);
 	}
 
 	public interface IHotelService : IEntityProvider<IHotel>
