--- src/TravelDbConnectionString
@@ -1,1 +1,0 @@
-User ID=test;Password=test;Host=db;Port=5432;Database=db_test_1;Pooling=true;
